+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * contributions for various models
 
+### Fixed
+* lexeme view now filters by language
+
 ## [0.0.8] - 2023-06-17
 
 ### Added