+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.9] - 2023-06-21
 
+### Fixed
+* missing dependency
+
 ### Added
 * contributions for various models
 * parts of speech for morphs