+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.9] - 2023-06-21
+## [0.0.10] - 2023-06-21
 
 ### Fixed
 * missing dependency
+
+## [0.0.9] - 2023-06-21
 
 ### Added
 * contributions for various models
@@ -95,7 +97,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Initial release
 
-[Unreleased]: https://github.com/fmatter/clld-morphology-plugin/compare/0.0.9...HEAD
+[Unreleased]: https://github.com/fmatter/clld-morphology-plugin/compare/0.0.10...HEAD
+[0.0.10]: https://github.com/fmatter/clld-morphology-plugin/compare/0.0.9...0.0.10
 [0.0.9]: https://github.com/fmatter/clld-morphology-plugin/compare/0.0.8...0.0.9
 [0.0.8]: https://github.com/fmatter/clld-morphology-plugin/compare/0.0.7...0.0.8
 [0.0.7]: https://github.com/fmatter/clld-morphology-plugin/compare/0.0.6...0.0.7